requirements.txt
fastapi
uvicorn
gunicorn
transformers
sentence-transformers
faiss-cpu
pandas
numpy
torch

